requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
fonttools==4.32.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
mecab-python3==1.0.5
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
plac==1.3.5
pycodestyle==2.8.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
tweepy==4.8.0
unidic==1.1.0
urllib3==1.26.9
wasabi==0.9.1
wordcloud==1.8.1
